requirements.txt
httpx
pyyaml
orjson